ai/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf
python-docx
easyocr
opencv-python-headless
numpy
httpx
pillow
